Fix out-of-bounds access in three_parts. Fix mishandling of non-null terminated string_views in jwt_set::case_compare. Address several conversion and unreachability warnings

+++ include/jwt/algorithm.hpp
@@ -358,7 +358,7 @@ struct HMACSign
 
     unsigned char* res = HMAC(Hasher{}(),
                               key.data(),
-                              key.length(),
+                              static_cast<int>(key.length()),
                               reinterpret_cast<const unsigned char*>(data.data()),
                               data.length(),
                               reinterpret_cast<unsigned char*>(&sign[0]),
